challenges/oracle_of_uhasselt/src/requirements.txt
pycryptodome==3.14.1
python-dotenv==0.19.2
telnetserver==0.5
